Changed all appropriate services to execute isInUse() before deleting the provided domain. If a domain is in use a DomainInUseException is thrown to the calling code (controller). The associated controller delete actions are updated to catch and display an error to the user if this occurs.

+++ grails-app/services/carm/application/ApplicationTypeService.groovy
@@ -4,6 +4,7 @@ import org.springframework.security.access.prepost.PostFilter
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.security.acls.domain.BasePermission
 import org.springframework.transaction.annotation.Transactional
+import carm.exceptions.DomainInUseException
 
 class ApplicationTypeService {
 
@@ -57,6 +58,10 @@ class ApplicationTypeService {
     @Transactional
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     void delete(ApplicationType applicationType) {
+        if (isInUse(applicationType)) {
+            throw new DomainInUseException()
+        }
+
         applicationType.delete()
     }
