+++ grails-app/services/carm/deployment/ModuleDeploymentTestStateService.groovy
@@ -4,6 +4,7 @@ import org.springframework.security.access.prepost.PostFilter
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.security.acls.domain.BasePermission
 import org.springframework.transaction.annotation.Transactional
+import carm.exceptions.DomainInUseException
 
 class ModuleDeploymentTestStateService {
 
@@ -57,6 +58,10 @@ class ModuleDeploymentTestStateService {
     @Transactional
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     void delete(ModuleDeploymentTestState moduleDeploymentTestState) {
+        if (isInUse(moduleDeploymentTestState)) {
+            throw new DomainInUseException()
+        }
+
         moduleDeploymentTestState.delete()
     }
