+++ grails-app/services/carm/project/ProjectCategoryService.groovy
@@ -3,6 +3,7 @@ package carm.project
 import org.springframework.transaction.annotation.Transactional
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.security.acls.domain.BasePermission
+import carm.exceptions.DomainInUseException
 
 class ProjectCategoryService {
 
@@ -56,6 +57,10 @@ class ProjectCategoryService {
     @Transactional
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     void delete(ProjectCategory projectCategory) {
+        if (isInUse(projectCategory)) {
+            throw new DomainInUseException()
+        }
+
         projectCategory.delete()
     }
