+++ grails-app/services/carm/release/ApplicationReleaseTestStateService.groovy
@@ -4,6 +4,7 @@ import org.springframework.security.access.prepost.PostFilter
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.security.acls.domain.BasePermission
 import org.springframework.transaction.annotation.Transactional
+import carm.exceptions.DomainInUseException
 
 class ApplicationReleaseTestStateService {
 
@@ -57,6 +58,10 @@ class ApplicationReleaseTestStateService {
     @Transactional
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     void delete(ApplicationReleaseTestState applicationReleaseTestState) {
+        if (isInUse(applicationReleaseTestState)) {
+            throw new DomainInUseException()
+        }
+
         applicationReleaseTestState.delete()
     }
