+++ grails-app/services/carm/sourcecontrol/SourceControlRoleService.groovy
@@ -5,6 +5,7 @@ import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.security.acls.domain.BasePermission
 import org.springframework.transaction.annotation.Transactional
 import carm.application.ApplicationRole
+import carm.exceptions.DomainInUseException
 
 class SourceControlRoleService {
 
@@ -58,6 +59,10 @@ class SourceControlRoleService {
     @Transactional
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     void delete(SourceControlRole sourceControlRole) {
+        if (isInUse(sourceControlRole)) {
+            throw new DomainInUseException()
+        }
+
         sourceControlRole.delete()
     }
