+++ grails-app/services/carm/sourcecontrol/SourceControlServerService.groovy
@@ -5,6 +5,7 @@ import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.security.acls.domain.BasePermission
 import org.springframework.transaction.annotation.Transactional
 import carm.application.Application
+import carm.exceptions.DomainInUseException
 
 class SourceControlServerService {
 
@@ -62,6 +63,10 @@ class SourceControlServerService {
     @Transactional
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     void delete(SourceControlServer sourceControlServer) {
+        if (isInUse(sourceControlServer)) {
+            throw new DomainInUseException()
+        }
+
         sourceControlServer.delete()
     }
